+++ app/src/main/java/com/example/adoteme_app/domain/repository/IRequisicaoRepository.kt
@@ -1,5 +1,5 @@
 package com.example.adoteme_app.domain.repository
 
 interface IRequisicaoRepository {
-    suspend fun solicitarAdocao(idAdotante: Int, idAnimal: Int): Boolean
+    suspend fun solicitarAdocao(idAdotante: Long, idAnimal: Long): Boolean
 }
