+++ app/src/main/java/com/example/adoteme_app/model/RequisicaoCreateDto.kt
@@ -1,6 +1,6 @@
 package com.example.adoteme_app.model
 
 data class RequisicaoCreateDto(
-    val idAdotante: Int,
-    val idAnimal: Int
+    val idAdotante: Long,
+    val idAnimal: Long
 )
